+++ NEWS.md
@@ -1,6 +1,7 @@
 # glamr 1.0.1
 * Add set of functions to store, access, and load Panorama credentials - `set_pano`,
- `pano_user()`, `pano_pwd()`, and `load_secrets()`.
+ `pano_user()`, `pano_pwd()`, `pano_session()`, and `load_secrets()`.
+* Add functions to discover and download output datasets from Panorama: `pano_content()`, `pano_elements()`, `pano_unpack()`, `pano_extract()`, `pano_download()`
 
 # glamr 1.0.0
 * Fixed `source_info()` to allow it to work correctly with FSD (misspelling in the filename).